requirements.txt
tensorflow
tensorflow-probability
trfl
dm-sonnet
ma-gym
